only use one environment variable for GitHub access token
Both release-it and lerna-changelog require a GitHub personal access token.
Both read it from an environment variable but by default not using the same
name. release-it defaults to `$GITHUB_TOKEN` while lerna-changelog defaults
to `$GITHUB_AUTH`. It could only be customized for release-it. Therefore
changing it to `$GITHUB_AUTH`.

Also documents that this environment variable is required. It wasn't
documented so far at all.

Background information could be found here:
- https://github.com/release-it/release-it/blob/master/docs/github-releases.md
- https://github.com/lerna/lerna-changelog#github-token