Avoid downgrading release-it / release-it-lerna-changelog.
When the project that we are runnign against has a newer version of
`release-it` or `release-it-lerna-changelog` we would always downgrade
their specified version to whatever we have specified in our own
`devDependencies`.

This is pretty bad behavior, and has been fixed. Now if the project
defines a dev dependency that is _newer_ than our range, we use it
directly.